replace plan with calendar, fix taskstore computed
