+++ src/views/HomeView.vue
@@ -1,25 +1,9 @@
 <script setup>
-import { ref, computed, watch, onMounted } from "vue";
-import { date } from "quasar";
-import apiClient from "../stores/api.client";
+import { useTaskStore } from "../stores/TaskStore";
 import GoalSelect from "../components/GoalSelect.vue";
 import WeekTimelineEntry from "../components/WeekTimelineEntry.vue";
 
-const GOAL_FILTER_KEY = "Goal-Filter";
-const current = date.formatDate(new Date(), "YYYY-w");
-const goal = ref(null);
-const url = computed(() =>
-  goal.value ? "/plan/?goal=" + goal.value : "/plan/"
-);
-const { data: plan, mutate } = apiClient.read(url);
-onMounted(() => {
-  const stored = localStorage.getItem(GOAL_FILTER_KEY);
-  if (stored) goal.value = parseInt(stored);
-});
-watch(goal, (data) => {
-  if (data) localStorage.setItem(GOAL_FILTER_KEY, data);
-  else localStorage.removeItem(GOAL_FILTER_KEY);
-});
+const store = useTaskStore();
 </script>
 
 <template>
@@ -27,20 +11,42 @@ watch(goal, (data) => {
     <q-toolbar>
       <q-toolbar-title>Plan</q-toolbar-title>
       <GoalSelect
-        v-model="goal"
+        v-model="store.filter.goal"
         label="For Goal"
         dense
         borderless
         class="col-6 col-sm-3 col-lg-2"
       />
+      <q-btn flat round icon="filter_list">
+        <q-menu>
+          <q-list>
+            <q-item>
+              <q-item-section>
+                <q-item-label>Hide passed done</q-item-label>
+              </q-item-section>
+              <q-item-section side>
+                <q-toggle v-model="store.filter.done" />
+              </q-item-section>
+            </q-item>
+            <q-item>
+              <q-item-section>
+                <q-item-label>Hide passed recurring</q-item-label>
+              </q-item-section>
+              <q-item-section side>
+                <q-toggle v-model="store.filter.recurring" />
+              </q-item-section>
+            </q-item>
+          </q-list>
+        </q-menu>
+      </q-btn>
     </q-toolbar>
-    <q-timeline v-if="plan" layout="dense" class="q-px-md">
-      <template v-for="(week, key) in plan" :key="key">
+    <q-timeline v-if="store.calendar" layout="dense" class="q-px-md">
+      <template v-for="(week, key) in store.calendar" :key="key">
         <WeekTimelineEntry
           v-if="week.day"
-          :color="key === current ? 'orange' : ''"
+          :color="key === store.currentWeek ? 'orange' : ''"
           :week="week"
-          @mutate="mutate"
+          @mutate="store.mutate"
         />
       </template>
     </q-timeline>
